Package renaming
Common

com.ludoscity.common->com.ludoscity.herdr.commmon

Android&iOS

toyed with the idea of having specific package/applicationId names and decided against for now
